REF: move DTI/TDI/PI get_value to ExtensionIndex, associated cleanups
